move back exception.hpp to exception.h
