+++ Core/interface/exception.h
