+++ Core/src/ConfigReader.cpp
@@ -12,7 +12,7 @@ This file is part of https://github.com/hh-italian-group/TauMLTools. */
 #include <functional>
 #include <boost/algorithm/string.hpp>
 
-#include "TauMLTools/Core/src/exception.hpp"
+#include "TauMLTools/Core/interface/exception.h"
 #include "TauMLTools/Core/interface/EnumNameMap.h"
 #include "TauMLTools/Core/interface/TextIO.h"
