+++ Core/src/EventIdentifier.cpp
@@ -6,7 +6,7 @@ This file is part of https://github.com/hh-italian-group/TauMLTools. */
 #include <sstream>
 #include <map>
 #include <boost/algorithm/string.hpp>
-#include "TauMLTools/Core/src/exception.hpp"
+#include "TauMLTools/Core/interface/exception.h"
 
 namespace analysis {
